// src/types/DW_Rest.d.ts
export interface ILinkModel {
  Links?: ILink[];
}

/**
 * Wraps all necessary logon properties
 *
 * @interface ILogonModel
 */
export interface ILogonModel {
  Username: string;
  Password: string;
  Organization: string;
  HostID: string;
  RedirectToMyselfInCaseOfError: boolean;
  RememberMe: boolean;
}

/**
 *The response after logon
 *
 * @interface ILogonResponse
 * @extends {ILinkModel}
 */
export interface ILogonResponse extends ILinkModel {
  Documentation: {
    Links: ILink[];
  };
}

/**
 * A link
 *
 * @interface ILink
 */
export interface ILink {
  rel: string;
  href: string;
  type: string;
}
// ...
export interface IFileCabinet extends ILinkModel {
  Name: string;
  Color: string;
  Id: string;
  IsBasket: boolean;
  Default: boolean;
  Documents: [IDocument];
}
// ...
export interface IDocument extends ILinkModel {
  Fields?: IDocumentIndexField[];
  Flags?: {
    IsCold: boolean;
    IsDBRecord: boolean;
    IsCheckedOut: boolean;
    IsCopyRightProtected: boolean;
    IsVoiceAvailable: boolean;
    HasAppendedDocuments: boolean;
    IsProtected: boolean;
    IsDeleted: boolean;
    IsEmail: boolean;
  };
  Sections?: ISection[];
  ContentType?: string;
  TotalPages?: number;
  Id?: number;
  Title?: string;
  LastModified?: string;
  LastModifiedSpecified?: boolean;
  CreatedAt?: string;
  CreatedAtSpecified?: boolean;
  FileSize?: number;
  SectionCount?: number;
  IntellixTrust?: string;
  VersionStatus?: string;
  FileCabinetId?: string;
  Version?: IDocumentVersion;
  ApplicationProperties?: IDocumentApplicationProperty[];
}

/**
 * DocuWare InputDocument Model
 *
 * @export
 * @interface IInputDocument
 * @extends {ILinkModel}
 */
export interface IInputDocument extends ILinkModel {
  fields?: IDocumentIndexField[];
  flags?: IInputFlags;
  sections?: IInputSection[];
  fileChunk?: IUploadedFileChunk;
}

/**
 * DocuWare DocumentIndexFieldValueBase Model
 *
 * @export
 * @interface IDocumentIndexFieldValueBase
 */
export interface IDocumentIndexFieldValueBase {
  item: any;
  itemElementName: ItemChoiceType;
}

/**
 * DocuWare DocumentIndexField Model
 *
 * @export
 * @interface IDocumentIndexField
 * @extends {IDocumentIndexFieldValueBase}
 */
export interface IDocumentIndexField extends IDocumentIndexFieldValueBase {
  pointAndShootInfo?: IPointAndShootInfo;
  fieldName: string;
  fieldLabel?: string;
  isNull?: boolean;
  readOnly?: boolean;
}

/**
 * DocuWare PointAndShootInfo Model
 *
 * @export
 * @interface IPointAndShootInfo
 */
export interface IPointAndShootInfo {
  box: IDWRectangle[];
  pageNumber: number;
}

/**
 * DocuWare DWRectangle Model
 *
 * @export
 * @interface IDWRectangle
 */
export interface IDWRectangle {
  left: number;
  top: number;
  width: number;
  height: number;
}
// ...
export interface ISection extends ILinkModel {
  Id: string;
  ContentType: string;
  HaveMorePages: boolean;
  PageCount: number;
  FileSize: number;
  OriginalFileName: string;
  ContentModified: string;
  Pages: ILinkModel;
}
// ...
export interface IFieldList {
  Field: IDocumentIndexField[];
}

/**
 * DocuWare Dialog Field Model
 *
 * @interface IDlgField
 * @extends {ILinkModel}
 */
export interface IDlgField extends ILinkModel {
  DBFieldName: string;
  DlgLabel: string;
  DWFieldType: ItemChoiceType;
  Visible: boolean;
  Length?: number;
  AllowExtendedSearch: boolean;
  AllowFiltering: boolean;
  AssignedInternalSelectList: boolean;
}
// ...
export interface IDocumentApplicationProperty {
  Name: string;
  Value: string;
}

/**
 * DocuWare Query Result
 *
 * @interface IDocumentsQueryResult
 * @extends {ILinkModel}
 */
export interface IDocumentsQueryResult extends ILinkModel {
  Count: number;
  Items: IDocument[];
  Next: string;
  Download: string;
}

/**
 * DocuWare Dialog Model
 *
 * @interface IDialog
 * @extends {ILinkModel}
 */
export interface IDialog extends ILinkModel {
  Id: string;
  FileCabinetId: string;
  FileCabinetName: string;
  DisplayName: string;
  IsForMobile: boolean;
  AssignedDialogId: string;
  Color: string;
  IsDefault: boolean;
  Type: DialogType;
  Fields: IDlgField[];
  Query: {
    Expression: string;
    Fields: string[];
    Links: ILink[];
  };
  Properties: {
    LogicalOperator: {
      DefaultOperator: Operation;
      Show: boolean;
    };
  };
}
// ...
export interface IDocumentVersion {
  Major: number;
  Minor: number;
}
// ...
const enum ItemChoiceType {
  Int = 'Int',
  DateTime = 'DateTime',
  String = 'String',
  Decimal = 'Decimal',
  Date = 'Date',
  Table = 'Table',
  Keywords = 'Keywords',
}

/**
 * DocuWare Dialog Types
 *
 * @enum {number}
 */
const enum DialogType {
  Search = 'Search',
  Store = 'Store',
  ResultList = 'ResultList',
  InfoDialog = 'InfoDialog',
  TaskList = 'TaskList',
}
// ...
const enum Operation {
  And = 'And',
  Or = 'or',
}

/**
 * Sort order
 *
 * @enum {number}
 */
const enum SortOrder {
  Asc = 'Asc',
  Desc = 'Desc',
}